requirements.txt
awscrt==0.14.7
awsiotsdk==1.11.9
certifi==2022.9.24
pip==22.2.2
setuptools==63.4.1
wheel==0.37.1
